refactor(assistant): extract AssistantConversation from AssistantDrawer
Move the conversation UI, state, and handlers out of AssistantDrawer into
a standalone AssistantConversation component. AssistantDrawer becomes a
thin Drawer wrapper around it, preserving the existing resource-scoped
behavior. This lets a future full-tab, non-drawer host reuse the same
conversation surface.
